+++ libraries/extractor/src/main/java/androidx/media3/extractor/DtsUtil.java
@@ -41,6 +41,17 @@ public final class DtsUtil {
   private static final int SYNC_VALUE_14B_BE = 0x1FFFE800;
   private static final int SYNC_VALUE_LE = 0xFE7F0180;
   private static final int SYNC_VALUE_14B_LE = 0xFF1F00E8;
+  /**
+   * DTS Extension Substream Syncword (in different Endianness). See ETSI TS 102 114 (V1.6.1)
+   * Section 7.4.1.
+   */
+  private static final int SYNC_EXT_SUB_LE = 0x25205864;
+  /**
+   * DTS FTOC Sync words (in different Endianness). See ETSI TS 103 491 (V1.2.1) Section 6.4.4.1.
+   */
+  private static final int SYNC_FTOC_LE = 0xF21B4140;
+
+  private static final int SYNC_FTOC_NON_SYNC_LE = 0xE842C471;
   private static final byte FIRST_BYTE_BE = (byte) (SYNC_VALUE_BE >>> 24);
   private static final byte FIRST_BYTE_14B_BE = (byte) (SYNC_VALUE_14B_BE >>> 24);
   private static final byte FIRST_BYTE_LE = (byte) (SYNC_VALUE_LE >>> 24);
@@ -149,6 +160,16 @@ public final class DtsUtil {
    * @return The number of audio samples represented by the syncframe.
    */
   public static int parseDtsAudioSampleCount(ByteBuffer buffer) {
+    if ((buffer.getInt(0) == SYNC_FTOC_LE) || (buffer.getInt(0) == SYNC_FTOC_NON_SYNC_LE)) {
+      // Check for DTS:X Profile 2 sync or non sync word and return 1024 if found. This is the only
+      // audio sample count that is used by DTS:X Streaming Encoder.
+      return 1024;
+    } else if (buffer.getInt(0) == SYNC_EXT_SUB_LE) {
+      // Check for DTS Express sync word and return 4096 if found. This is the only audio sample
+      // count that is used by DTS Streaming Encoder.
+      return 4096;
+    }
+
     // See ETSI TS 102 114 subsection 5.4.1.
     int position = buffer.position();
     int nblks;